fix: move proposer client config
